+++ backend/CrispyOctoChainsaw/CrispyOctoChainsaw.API/Controllers/SystemAdminsController.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using CrispyOctoChainsaw.API.Contracts;
-using CrispyOctoChainsaw.Domain;
+using CrispyOctoChainsaw.Domain.Model;
 using CrispyOctoChainsaw.Domain.Interfaces;
 using Microsoft.AspNetCore.Mvc;
