+++ backend/CrispyOctoChainsaw/CrispyOctoChainsaw.DataAccess.Postgres/DataAccessMappingProfile.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using CrispyOctoChainsaw.DataAccess.Postgres.Entities;
-using CrispyOctoChainsaw.Domain;
+using CrispyOctoChainsaw.Domain.Model;
 
 namespace CrispyOctoChainsaw.DataAccess.Postgres
 {
