+++ backend/CrispyOctoChainsaw/CrispyOctoChainsaw.Domain/Model/User.cs
@@ -1,27 +1,48 @@
-﻿using CSharpFunctionalExtensions;
+﻿using System.Net.Mail;
+using CSharpFunctionalExtensions;
 
 namespace CrispyOctoChainsaw.Domain.Model
 {
     public record User
     {
-        public string Id { get; }
+        public Guid Id { get; }
+
+        public string Email { get; }
 
         public string UserName { get; }
 
-        private User(string id, string userName)
+        private User(Guid id, string email, string userName)
         {
             Id = id;
+            Email = email;
             UserName = userName;
         }
 
-        public static Result<User> Create(string userName)
+        public static Result<User> Create(string email, string userName)
         {
             if (string.IsNullOrWhiteSpace(userName))
             {
-                return Result.Failure<User>("Имя пользователя не может быть пустым");
+                return Result.Failure<User>("UserName cannot be empty");
+            }
+            if (IsValidEmail(email) == false)
+            {
+                return Result.Failure<User>("Email is incorrect");
             }
 
-            return new User(string.Empty, userName);
+            return new User(Guid.Empty, email, userName);
+        }
+
+        private static bool IsValidEmail(string email)
+        {
+            try
+            {
+                var address = new MailAddress(email);
+                return address.Address == email;
+            }
+            catch
+            {
+                return false;
+            }
         }
     }
 }
